+++ src/test/kotlin/lotto/LottoScannerTest.kt
@@ -1,40 +1,15 @@
 package lotto
 
+import lotto.model.Lotto
+import lotto.model.LottoNumber
+import lotto.model.LottoScanner
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-
-class LottoResult(
-    private val ranks: List<Rank>,
-) {
-    fun calculateProfit(): Double {
-        val totalWinningMoney = ranks.sumOf { it.winningMoney }
-        return totalWinningMoney / (ranks.size * 1_000).toDouble()
-    }
-}
-
-class LottoScanner(
-    private val winningNumbers: Lotto,
-) {
-    fun getResult(lottos: List<Lotto>): LottoResult {
-        val result = lottos.map { getRank(it) }
-        return LottoResult(result)
-    }
-
-    fun getRank(lotto: Lotto): Rank {
-        val countOfMatch: Int = getCountOfMatch(lotto)
-        val matchBonus: Boolean = getMatchBonus(lotto)
-        return Rank.valueOf(countOfMatch, matchBonus)
-    }
-
-    fun getCountOfMatch(lotto: Lotto): Int = winningNumbers.getNumbers().intersect(lotto.getNumbers().toSet()).size
-
-    fun getMatchBonus(lotto: Lotto): Boolean = lotto.getNumbers().contains(winningNumbers.getBonusNumber())
-}
 
 class LottoScannerTest {
     @Test
     fun `로또 당첨을 판단한다`() {
-        val bonusNumber: LottoNumber = LottoNumber(7)
+        val bonusNumber = LottoNumber(7)
 
         val winningNumbers =
             Lotto(
@@ -46,6 +21,7 @@ class LottoScannerTest {
                     LottoNumber(5),
                     LottoNumber(6),
                 ),
+                bonusNumber,
             )
 
         val lottoTicket =
