+++ src/test/kotlin/lotto/LottoTest.kt
@@ -1,34 +1,10 @@
 package lotto
 
+import lotto.model.Lotto
+import lotto.model.LottoNumber
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-
-class Lotto(
-    private val numbers: List<LottoNumber> = generateLotto(),
-    private val bonusNumber: LottoNumber? = null,
-) {
-    init {
-        require(numbers.size == 6)
-        require(numbers.size == numbers.toSet().size)
-        require(!numbers.contains(bonusNumber))
-    }
-
-    fun getSize() = numbers.size
-
-    fun getNumbers() = numbers.toList()
-
-    fun getBonusNumber() = bonusNumber
-
-    companion object {
-        private fun generateLotto(): List<LottoNumber> =
-            (1..45)
-                .shuffled()
-                .take(6)
-                .sorted()
-                .map { LottoNumber(it) }
-    }
-}
 
 class LottoTest {
     @Test
